ArrayTransformer 코드 이동
